requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Authlib==1.3.1
Flask-Login==0.6.3
